remove hir module
