+++ ogma/src/eng/blk.rs
@@ -18,6 +18,16 @@ impl<'a> Block<'a> {
             .1
     }
 
+    /// The input [`Type`] of the block.
+    pub fn in_ty(&self) -> &Type {
+        &self.in_ty
+    }
+
+    /// Returns the _number of arguments remaining_.
+    pub fn args_len(&self) -> usize {
+        self.args.len()
+    }
+
     /// Assert that this block will return the given type.
     ///
     /// Asserting an output type gives the type inferer knowledge about this block's output.
@@ -41,6 +51,23 @@ impl<'a> Block<'a> {
         self.chgs
             .push(graphs::tygraph::Chg::KnownOutput(self.node.idx(), ty).into());
     }
+
+    /// Gets the flag that matches a given name.
+    ///
+    /// If no name is given with `None`, _the first flag first is returned, if there is one._
+    ///
+    /// > The flag is **removed** from the flag stack.
+    pub fn get_flag<'b, N: Into<Option<&'b str>>>(&mut self, flag: N) -> Option<Tag> {
+        match flag.into() {
+            Some(name) => self
+                .flags
+                .iter()
+                .position(|x| x.str() == name)
+                .map(|i| self.flags.remove(i)),
+            None => self.flags.pop(),
+        }
+    }
+
 
     /// See if there is a next argument node, without popping off the stack.
     pub fn peek_next_arg_node(&self) -> Option<graphs::ArgNode> {
@@ -122,5 +149,46 @@ impl<'a> Block<'a> {
             .peek_next_arg_node()
             .ok_or_else(|| Error::insufficient_args(self.op_tag(), self.args_count, None))?;
         self.inject_manual_var_into_arg_locals(n, name, ty)
+    }
+}
+
+/// Evalulation functions.
+impl<'a> Block<'a> {
+    /// Most flexible evaluation option, but also most brittle.
+    ///
+    /// **BE EXTRA CAREFUL WITH THE `out_ty` THAT IT MATCHES THE EVAL VALUE.
+    /// It is recommended to thoroughly test code paths through here to ensure type validity.**
+    ///
+    /// # Usage
+    /// - Input value ([`Value`]) should be cast to expected input type (use `.try_into()?`).
+    /// - Arguments can use this type if they are expecting the blocks input.
+    ///
+    /// ## Arguments
+    /// Arguments **do not need to use blocks input**. If no input type is needed, the argument can
+    /// be built with `Type::Nil` and `Value::Nil` can be passed on through!
+    pub fn eval<F>(self, out_ty: Type, f: F) -> Result<Step>
+    where
+        F: Fn(Value, Context) -> StepR,
+        F: Func<StepR>,
+    {
+        self.finalise(&out_ty)?;
+        Ok(Step {
+            out_ty,
+            f: Arc::new(f),
+        })
+    }
+
+    /// Preferred way of creating a eval step.
+    ///
+    /// This supplies the [`Value`] input but uses type inference on `O` to get the output type.
+    pub fn eval_o<F, O>(self, f: F) -> Result<Step>
+    where
+        F: Fn(Value, Context) -> Result<(O, Environment)>,
+        F: Func<Result<(O, Environment)>>,
+        O: AsType + Into<Value>,
+    {
+        self.eval(O::as_type(), move |v, c| {
+            f(v, c).map(|(x, e)| (Into::into(x), e))
+        })
     }
 }
